bitops: added bitops test_bit function.
Added the first bit ops function: test_bit. This is a generic
implementation (i.e. architectures can override it by a faster
function). The function is defined as follows:

Test the value of a given bit in a given flags. The result of the bit
and only that bit will be returned in the 0-bit position. The return
value of this function can therefore only be 0 or 1.

Signed-off-by: Michel Megens <dev@michelmegens.net>