avr: simul: added output registers for avrsimul.
Added an output and input register for the AVR emulator `avrsimul'.
These registers are within the I/O memory, but unused by the MCU.
Writing and reading from these registers results in interaction with the
user of avrsimul. These registers will be directly bound to the in- and
output functions of libc.

Note: these registers should NEVER be used on real hardware, since the
behaviour of unused I/O memory is undefined (usually).

Signed-off-by: Michel Megens <dev@michelmegens.net>