kbuild: app: added support to build out of tree.
Applications can from now on be build out of tree (i.e. in another
directory). This is done by invoking the following make command:

  make -C $(etaos-src) A=$(PWD) ARCH=$(arch) CROSS_COMPILE=$(cross) app

See the documentation at etaos.michelmegens.net for an example Makefile
and Kbuild file for building applications this way. This way is also the
preferred and recommended way.

Signed-off-by: Michel Megens <dev@michelmegens.net>