sync: mutex: added initial mutex implementations.
Added the first synchronization principle: mutexes. If event mutexes are
not configured for building, mutexes are compiled as spinlocks. If a
lock is requested it will test the value of the lock and if its 0, set
it to 1 and return control to the application. If the value is NOT 0 it
will keep spinning in a loop testing it, over and over again until the
value of the lock becomes zero.

Signed-off-by: Michel Megens <dev@michelmegens.net>