mm: added an initial version for mm.
Implemented an initial version of the memeory allocator (mm). The
allocation API consists of a generic structure driver and an allocation
algorithm driving (i.e. using) the structure driver for memory
allocation purposes.

An example of a structure driver is found in mm/mlist.c. This file
contains generic functions to alter a sorted linked list containing all
free memory nodes.

The only allocator that has been implemented in this patch is the
first-fit allocator. Another allocator is on the way, the best-fit
allocator. The allocation algorithm is going to use the same storage
principles as the first-fit algorithm.

Signed-off-by: Michel Megens <dev@michelmegens.net>