libc: string: added a string maniplation module.
lib/string contains the '<string.h>' part of libc. These functions are
all, in some way, relevant to C style strings (char *). The functions
implemented in v1 are:

	* strcmp (compare 2 strings)
	* strchr (find first occurance of a character)
	* strlen (find the length of a string)
	* strnlen (see if a string is longer than a given amount)

The return value of strnlen is the value of strlen if the string is NOT
longer than the given amount, if it is the given length is returned.

Signed-off-by: Michel Megens <dev@michelmegens.net>