atomic: added atomic types and API funcs.
Added the types:

	* atomic_t;
	* atomic64_t;

The atomic_t type is an integer type with the length `signed int'. The
atomic64_t type is defined as an `int64_t' type and is for that reason
ALWAYS 64 bits long. Usage of the latter should be reduced for data
space reasons only.

The currently implemented atomic API functions:

	* atomic_inc & atomic64_inc
	* atomic_get & atomic64_get

The get functions return their respective types mentioned above.

Signed-off-by: Michel Megens <dev@michelmegens.net>