time: added timer and clocksource structures.
The timer structure is the frontend structure for a timer. It defines
how often it fires, and what to do when it fires. You can compare this
structure to the pointers on the clock. Behind the clock there's an
engine which defines how fast these pointers are being moved. It is
basically the driver of the clock. In ETA/OS there is a specific
structure to define this mechanism: the clocksource.

Signed-off-by: Michel Megens <dev@michelmegens.net>