irq: chip: fixed a list_add bug.
The parameter order of list_add is to first pass the new object you want
to add to the list and secondly the list head. The list head is the
object the where the first object will be added after. Consider the
following list:

X - Y - Z

If you now call list_add as follows:
	* list_add(&A, &Y)

The list will look like:

X - Y - A - Z

The list_add in irq_chip_add_irq called list_add in the wrong way (i.e.
the head first and the new object as second parameter). Flipped them to
fix.

Signed-off-by: Michel Megens <dev@michelmegens.net>