dev: core: rewritten device create
device_create has been rewritten to a new template. Device create will,
from now on take a name, a private data pointer and a pointer to the
appropriate file operations.

device_initialize is the cousin of device create, but it doesn't
allocate a new device. It takes one which is already allocated in its
parameters and initializes it.

Signed-off-by: Michel Megens <dev@michelmegens.net>