usr: tests: added automated tests to usr/
Automated tests can be added to usr/ from now on. These tests can be
designed to run on real hardware or on simulators of real hardware. To
run all tests issue the following command in usr/

`make tests'

To run a specific test, cd into the usr/tests directory and run:

`make <testname>'

Signed-off-by: Michel Megens <dev@michelmegens.net>