avr: timer: implemented the avr sys-clk init.
The AVR system clock is run of the AVR core timer 0, which is an 8 bit
HR timer. The timer is configured to interrupt on overflow, which occurs
at each time the timer hits 125 or 250 (depending on the CPU frequency). The
timer only works for CPU's running on either 8MHz or 16Mhz:

 8Mhz/64/125 = 1000Hz
16Mhz/64/250 = 1000Hz

Signed-off-by: Michel Megens <dev@michelmegens.net>