link: irq: fixed a linking error.
The interrupt vectors are defined from PROGMEM address 0x0 - 0xXXXX.
There was some linking error which placed garbage at the first ~30-ish
bytes of the program memory. Up until now, this wasn't a problem,
however as soon as interrupts come in play it caused a crash because the
right IRQ vector couldn't be found.

Signed-off-by: Michel Megens <dev@michelmegens.net>