sched: core: fixed a potential RQ bug
The RQ pointer in the `struct thread' data structure (struct thread::rq)
was potentially being violated by kill/wake queue editing functions.
Fixed by acquiring the current run queue as their destination run queue.

If that isn't their current RQ it will be removed from their current run
queue and reassigned to the run queue of the executing CPU. This might
introduce some thread bouncing on SMP systems, but the overhead is
minimal.

The 'currently running thread' is never violated and moved to another
run queue, because it already is on the run queue of the executing CPU.

Signed-off-by: Michel Megens <dev@michelmegens.net>