avr: init: fixed a big memory related bug.
When not using `kzalloc' you MUST initialise EVERY data field of a
structure, or it will be undefined. This happend in the thread creation
process and caused a nasty bug in the scheduler.

Fix:
* Implemented memset in the standard string library (TODO: optimize).
* Implemented kzalloc in include/mem.h

`kzalloc' uses memset if the stardard string library is being build, if
not, it uses its own (slow) algorithm to zero out the allocated memory.

Signed-off-by: Michel Megens <dev@michelmegens.net>