sched: evm: added support for event queue mutexes.
Event queue mutexes are mutexes with more overhead on locking and
unlocking. Whats the point, you might wonder. The point of using them
over spinlocks, is that they give up control over the CPU when they
enter a locking state. In other words, it is taking 'aquiring a
deadlock' to a whole new level, which is a good thing. It also frees up
CPU time for other threads, which would otherwise be waiting for a
thread doing nothing but waiting on a lock to get released.

Signed-off-by: Michel Megens <dev@michelmegens.net>