sched: core: added the __hot attrib to rq_schedule
GCC has quite allot of function attributes which can be used to optimize
the runtime performance of functions. One of these attributes is the
'__hot__' attribute. It is an attribute which is supposed to passed to
functions which are the hotspots of a program or system. In our case
`schedule' `rq_schedule' and `rq_switch_context' are a few of these hot
spots.

The attribute is warped by a definition in compiler-gcc.h as:

__hot => __attribute__((__hot__))

compiler.h contains a dummy definition in case ETA/OS is not being
compiled by GCC or without GCC extensions enabled.

Signed-off-by: Michel Megens <dev@michelmegens.net>