sched: core: fixed a rq_list_remove bug.
rq_list_remove was bugged if the entry we were looking for wasn't the
first entry in the list. The carriage variable was never set to the next
item in the list. Fixed by adding:

carriage = carriage->rq_next;

irq: thread: added support for threaded IRQ's.

The API to created threads ran in threaded context already existed, but
it wasn't implemented yet. This commit adds an implementation to the
API.

Signed-off-by: Michel Megens <dev@michelmegens.net>