sched: core: removed the some usages of current.
current(rq) is an internal definition to quickly get the currently
running thread on a run queue in the scheduling core. It was however
being used when there was already a local var initialised to the current
thread. Therefore, the uses of `current' have been removed where using
the reference variable is sufficient.

sched: rr: doc: Fixed a doc error on the sys_sched_class.

The system sched class of the round robin scheduling did also show up in
the generic sched::core documentation.

gpio: doc: added documentation to the gpio header.

Signed-off-by: Michel Megens <dev@michelmegens.net>