lib: stdio: added getc, fgetc and ioctl.
Implmented the functions:

	* getc
	* fgetc
	* ioctl

ioctl can be used to control several options of a device file. The
options you can control differ for each device. Please check the help
documents for the specific device driver to see what can be controlled
using ioctl.

getc and fgetc are both stream readers, used to read a single byte from
a stream. Please not that the STREAM_READ_FLAG has to be set in the
`struct file::flags' register.

Signed-off-by: Michel Megens <dev@michelmegens.net>