irq: threaded: fixed some stack issue's
Fixed some stack usage issue's on threaded IRQ's. The The wake up
function for threaded IRQ's was using a rather big portion of the stack.
Removed some redundant code and cleaned the function up in general to
reduce the stack usage. Moved the function in rq_schedule down into the
resched loop and under the run queue lock (where it should be).

Signed-off-by: Michel Megens <dev@michelmegens.net>