timer: count: added clock overflow protection
Added protection for counter overflows, and more specifically system
clock overflows. If the system clock would happen to overflow, the
entire system would crash. With this patch in place, the system will
detect a clock overflow silently and handle it peacefully to ensure that
the system will keep working as intended.

Signed-off-by: Michel Megens <dev@michelmegens.net>