time: fixed up time_after and time_before
Fixed time_after and time_before to work with the new system tick type.
Since the new system tick can be any width and either be signed or
unsigned, the old

	y - x < 0

trick won't work anymore to see if the system tick as advanced a certain
point. Fixed it by redefining time_after to

	x > 0

The way time_after and time_before are supposed to be called hasn't
changed (i.e. the API is compatible).

Signed-off-by: Michel Megens <dev@michelmegens.net>