avr: atmega1280: made a setup for the ATmega1280.
First preperations are made for the ATmega1280 architecture in this
patch.

Added in this update:

	* link scripts for the AVR51 sub architecture
	* I/O header (iom1280.h)
	* Makefile definitions (defconf/atmega1280.dc)

The ATmega328(p) have been moved to the appropriate defconf file
(atmega328.dc).

For the AVR architecture it is also no longer to define the amount of
available interrupt lines in the Kbuild configuration. The amount of
available IRQ lines is now dictated by the appropriate architecture
header (iomXXXX.h).

Signed-off-by: Michel Megens <dev@michelmegens.net>