mm: best-fit: added a basis for a bf allocator
A SLOB allocator can be implemented in roughly two ways; best-fit and
first-fit. The first-fit algorithm is the simplest and fastest
algorithm, but it lacks protection against fragmentation. The best-fit
algorithm is a little bit slower, but it tries to ensure that the best
fitting memory region is returned to the user requesting the region.

Signed-off-by: Michel Megens <dev@michelmegens.net>