spi: atmega: implemented the atmega SPI bus
Implemented the system SPI bus of the ATmega series microcontrollers.
This bus is initialised as a device instead of module, since it relies
on the GPIO subsystem to be fully initialised (which is a subystem (the
generic lib) and a module (the GPIO drivers)). Devices is the init level
after module.

Because it is initialised during the module setup SPI devices drivers
can not use the automatic system initialisation. This is an issue to
should be patched in the future.

Signed-off-by: Michel Megens <dev@michelmegens.net>