sched: core: rewritten `rq_schedule(rq)'
The core function rq_schedule has been rewritten to provide
better readability. The length of the function has been slightly
decreased and unnecessary `#ifdefs' have been removed.

The awakening of threads in EVM's queue's has been removed from
`rq_schedule()'. The code to provide the awakening has been moved to
a separate function, with only a call added to `rq_schedule()'.

Some functions which belong to the scheduling core, but were implemented
in other parts of the kernel have been moved to the appropriate place in
kernel/sched/core.c.

Signed-off-by: Michel Megens <dev@michelmegens.net>