sched: lottery: added a setup for the lottery algo
Lottery scheduling is a probabilistic scheduling algorithm. Each thread
is given a number of tickets ('lottery' tickets), and the scheduler
draws a random ticket when a new thread has to be select for execution.

The distribution of tickets does not need to be uniform; it is fine to
grant a thread more tickets then another. This results in probabilistic
priorities. The more tickets a thread has, the higher the chance that it
will be selected by the scheduler.

Signed-off-by: Michel Megens <dev@michelmegens.net>