sched: rr: fixed an issue on dyn priorities ranges
The dynamic priority updater in the round robin scheduler now takes a
maximum dynamic priority in account (in relation to the actual
priority).

The true priority is calculated by substracting the dynamic priority
from the user priority; therefore the dynamic priority may be equal to
the user priority at most (resulting in a true priority of 0).

Signed-off-by: Michel Megens <dev@michelmegens.net>