mm: stdlib: moved stdlib mm functions
Moved all crt (C runtime library) functions to the appropriate place
within the ETA/OS source (lib/).

Also a new crt function has been implemented; kcalloc. kcalloc tries to
allocate memory for an array. Its arguments are the number of elements
the array is ment to hold and the size of each element.

Signed-off-by: Michel Megens <dev@michelmegens.net>