sched: core: implemented thread_queue_wait.
thread_queue_wait is an alternative to sleep(). The difference is that
threads put to sleep using `sleep()' are basically 'dangeling' objects.
They are not in any queue. The thread_queue_wait function allows you to
let threads sleep in a thread queue.

These thread wait queue's are now also used in dev_sync_lock. Using
thread_queue_wait instead of sleep() in dev_sync_lock fixes a bug with
with the reinsertion of threads.

Signed-off-by: Michel Megens <dev@michelmegens.net>