sched: core: moved the dynamic prio update up.
cpu_schedule(cpu:int) is responsible for updating the dynamic
priorities of threads. This was done as one of the last functions
called. The only issue is that the run queue lock had to be locked
again when doing so.

This step, of locking the RQ (again), has been removed by moving
it up in the reschedule loop construction.

Signed-off-by: Michel Megens <dev@michelmegens.net>