dev: core: added dev_sync_wait to support multiops
dev_sync_wait does the same thing as dev_sync_lock, without unlocking
the actual sync mutex. This is useful in functions which have to do
multiple I/O operations, to the same device, in a row. A good example
would be the 24C02 (EEPROM) page write.

If you'd like to write 18 bytes to EEPROM you'll have to write 2 pages
(8 bytes each) and 1/4th of a page (2 bytes). As you can see now, this
requires 3 write operations, and every single write operation requires
the a timeout to allow the EEPROM chip to process data internally.
dev_sync_wait can be used in between these operations to force the
application to wait (asynchronously).

Signed-off-by: Michel Megens <dev@michelmegens.net>