eeprom: 24c02: implemented sequential reads/writes
The 24C02 now supports the stdio read/write calls by implementing a
sequential read and write. The write however, does not write in a single
sequence though, if the amount of bytes to be written is more then 8.

This is, because the chip works with page writes in sequential mode and
a single page is 8 bytes long. Therefore, the data will be split up in
pages + a left over and they will be written to the EEPROM chip in
seperate IO operations.

The sequential read does allow reading the full length of the EEPROM
chip in a single IO operation.

Signed-off-by: Michel Megens <dev@michelmegens.net>