sched: irq: fixed wake up process of threaded irqs
Previously, threaded IRQ's were using a thread queue to wait on after
the interrupt has been handled. IRQ's were being called while they were
not supposded to be called due to the the inherentance of the event
counter (i.e. left over 'ticks' are shipped over to the next one waiting
in line).

This issue has been fixed by removing the waiting queue all together.
The wake up function, irq_signal_threads, now loops through all
interrupts on the irq_chip of the CPU. Once a threaded IRQ has been
founded, it checks if events have been posted and if so, the threaded
that belongs to the IRQ is woken up.

Signed-off-by: Michel Megens <dev@michelmegens.net>