sched: core: added preemption time slice reset
After a new thread has been selected for execution, its time
slice should be reset to the default. This way the scheduler
remains fair. All threads should have the same time to execute
a thread.

Please note, that the time slice of the NEW thread is reset, not
the thread which is being scheduled out. Therefore, the run queue
may contain threads with no time left on their slice.

Signed-off-by: Michel Megens <dev@michelmegens.net>