kconfig: crurom: implemented crurom in kbuild
Kbuild is now capable of handling CRUROM files. To include local
files in the ROMFS add the following lines to the Kbuild file of
an application:

	crurom-y := dir1 dir2 .. dirN
	crurom-obj := romfsname.o

The `crurom-y' file lists all directory's to search for files. Please
note that these directory's will be searced recursively. The
`crurom-obj' variable sets a name for an object file to contain the
compiled romfs. This file will be made up of completely generated
content.

Signed-off-by: Michel Megens <dev@michelmegens.net>