eeprom: removed the use of internal index pointers
EEPROM chips no longer use an internal index (or offset) pointer
to indicate the internal read/write address. From now on the file
offset pointer is used. This value is dynamically changed after each
read or write operation. If manual manupulation of this value is
required, lseek can be used.

Signed-off-by: Michel Megens <dev@michelmegens.net>