vfs: open/close: added function pointer calls
The CRT open and close functions now also check if the file they are
opening or closing has a valid open or close function pointer. If so,
this function will be called before any critical data is assigned or
removed from the file.

The EEPROM and SRAM drivers now use this functionality to lock a device
mutex.

Signed-off-by: Michel Megens <dev@michelmegens.net>