sched: thread: updated thread name storage.
From now on, when allocating a new thread, the space to store the name
is automatically allocated. The maxiumum length of a thread name is 8
bytes (i.e. 8 characters (9, including the NULL terminator)).

When thread_create is invoked, the name given in the parameter list
will be copied into the thread structure. When a name is longer than
8 bytes the behavious is undefined.

Signed-off-by: Michel Megens <dev@michelmegens.net>