dev: init: added a chip level init
Chip level inits are now the lowest init level (i.e. 'chip' device
drivers are initialised last). This init level can be used on chips that
are waiting for a device to finish its initialisation (SRAM chips for
example; they need the SPI bus, which is a device level driver).

Signed-off-by: Michel Megens <dev@michelmegens.net>