romfs: implemented the uROM filesystem
Verry simple, read-only file system. ROMFS uses the program memory to
store a file (this file is built into the application during the
compilation process; this is done by the crurom tool). The filesystem
driver then adds a file during boot time for each entry that was created
by crurom. Only files are put into the ROM, directory names are included
in the file name.

The data will be copied into a file buffer when the file is opened. The
close operation will free this buffer again.

Files can be read by applications using the regular crt file API.

Signed-off-by: Michel Megens <dev@michelmegens.net>