romfs: added EOF return when index >= length.
When a file has been read up to its maximum length, an EOF is returned.

Also, a bug in the romfs_read has been fixed. The wrong file buffer
address was being passed to memcpy, resulting in faulty readings when a
file is read in multiple stages (i.e. not the entire file is read in one
operation).

Signed-off-by: Michel Megens <dev@michelmegens.net>