avr: crt: added program string printing
The CRT environment for the AVR architecture now supports the use of
program strings as format string for *printf functions. Example:

	printf_P(PSTR("Integer %s %i\n"), "example: ", 5);

The result on stdout will be:

	`Integer example: 5'

Do note the use PSTR. This preprocessor macro is used to convert a
normal string to a program memory string on systems that support it. If
this functionality is not supported (i.e. the architecture is not a
HARVARD architecture), the normal crt functions will be used (i.e.
printf instead of printf_P). This conversion is transparant.

Signed-off-by: Michel Megens <dev@michelmegens.net>