sched: core: optimisation of scheduling core
The scheduling core has received an optimisation iteration this patch.
An additional flag has been introduced to introduce a faster (but
slightly more complicated) reschedule policy (i.e. decision).

If a thread is supposed to be preempted by the system the
PREEMPT_NEED_RESCHED_FLAG is set. When the scheduler is ran, this thread
is checked together with the traditional THREAD_NEED_RESCHED_FLAG. If
only PREEMPT_NEED_RESCHED_FLAG is set and the 'next' thread is the idle
thread, the time slice of the 'current' thread is extended. If the
THREAD_NEED_RESCHED_FLAG is set, the thread will always be scheduled
out.

sched_cpu_to_rq has been renamed to cpu_to_rq to increase the ease of
reading.

Signed-off-by: Michel Megens <dev@michelmegens.net>