irq: added irq flag check
`irq_disabled()' can be used to check if any interrupts are disabled at
the moment of calling the function. If *any* interrupt is enabled at
that point the function will return false, if they are all disabled (or
the global interrupt flag in not enabled) the interrupt will return
true.

Signed-off-by: Michel Megens <dev@michelmegens.net>