sched: core: implemented IRQ preemption
If the CONFIG_PREEMPTION flag is enabled during the configuration of
ETA/OS, the system will start preempting threads using an interrupt.
Only if both preemption and the global interrupts are enbaled the thread
will be preempted. If the thread hasn't yet used up its entire time
slice the thread won't be preempted either.

Signed-off-by: Michel Megens <dev@michelmegens.net>