sched: core: move preemption call to sys tick IRQ
Move the `preempt_schedule()' call to the system independant system tick
IRQ handler.

The `preemptible()' function has been updated to also return true
whenever the interrupts are disabled AND the CPU is in IRQ context
(before, the interrupt context didn't matter and interrupts had to be
enabled in order to allow preemption).

Signed-off-by: Michel Megens <dev@bietje.net>