doc: sched-core: update docs for __schedule_need_resched
Updated the documentation for `__schedule_need_resched()' to better
explain the internal working of the reschedule algorithm / decision.

As per ETA/OS 4.4.1 the `__schedule_need_resched()' function decides
what to do based on whether or not the next thread is the idle thread.
If it is, and the thread is being preempted, the timeslice is reset and
no reschedule operation occurs. Otherwise the return value is the result
of the logical OR between the PREEMPT_NEED_RESCHED_FLAG and the
THREAD_NEED_RESCHED_FLAG.

Signed-off-by: Michel Megens <dev@bietje.net>