sched: profiler: add a scheduling profiler
Add a benchmark to measure the performance of the scheduler on a scope.
Two threads alternate the CPU as fast as possible. The main thread sets
a pin to a logical 1 and the second thread sets it to a logical zero.
The resulting signal is a decent indication for scheduling performance.

Signed-off-by: Michel Megens <dev@bietje.net>