sched: core: include 'next' thread in `preempt_chk'
It is the job of the scheduling core to let the scheduling algorithm
figure out the next thread to run. Once. It is wrong to first let the
scheduling core do it, and then do it again in preempt_chk. This could
cause issue's if the run queue changed (for whatever reason) in the mean
time.

Signed-off-by: Michel Megens <dev@bietje.net>