sched: core: optimise `__schedule_need_resched()'
Implement an optimisation in `__schedule_need_resched()': whenever the
THREAD_NEED_RESCHED_FLAG is set, the scheduler should reschedule.
Knowing this, it is always best to check this flag first and return
straight away if it is set.

Thereafter, the special case is checked: check if both preempt flag is
set and if the next thread is the idle thread. If either is false,
return the value of the PREEMPT_NEED_RESCHED_FLAG.

Signed-off-by: Michel Megens <dev@bietje.net>