sched: core: implement preemption mode PREEMT_FULL
Implement the new PREEMPT_FULL preemption mode. In this mode, the kernel
will preempt threads actively using the timer interrupt. Preemption will
not happen in the timer interrupt though, instead the timer interrupt
adds the address of the scheduler to the stack on TOP of the 'real' return
address. This way, when the interrupt returns it will first run the
scheduler.

Other interrupts can utilize this too by using the framework defined in
<asm/asmdef.h>.

Signed-off-by: Michel Megens <dev@bietje.net>