sched: core: fix concurrency and preemption issue's
Reset the PREEMPT_NEED_RESCHED_FLAG whenever a schedule is about to
occur, since leaving it on will give the thread an unfair start the next
time it is ran.

Also, when editting the timer lists on clocksources, the interrupts
should be disabled to prevent the timer interrupt from altering the list
at the 'same time'.

Signed-off-by: Michel Megens <dev@bietje.net>