sched: core: implement preemption after i2c and twi
The scheduler will attempt to preempt the current thread after the I2C
and SPI interrupts (for the ATmega architecture).

In addition to this, the core scheduling function, `__schedule()', now
runs the scheduling clock by itself, instead of relying on other sources
to run the timers. This means that applications that do not use the
scheduler, but do use timers need to invoke the timer processing
function on a regular basis.

Signed-off-by: Michel Megens <dev@bietje.net>