sched: core: remove `irq_thread_queue' definition
The scheduling core uses modified version of `signal()' and `wait()' to
handle the notification and waiting of IRQ threads
(`irq_thread_signal()' and `irq_thread_wait()' respectively). This was
changed in ETA/OS 4.6.1. However, the thread queue that was used to put
the IRQ threads on while they were idle was never removed.

Signed-off-by: Michel Megens <dev@bietje.net>