sched: core: only `wait()' if all events are handled
An IRQ thread should only be put in a waiting state using
`irq_thread_wait()` when all events on this IRQ are handled. The events
are tracked using a counter in `struct irq_thread_data::event_cnt`. When
this counter is zero the IRQ thread can safely go to sleep.

This policy will ensure that IRQ received while handling a threaded IRQ
are also handled immediately.

Signed-off-by: Michel Megens <dev@bietje.net>