sched: core: allow IRQ threads to be preemtive
Threaded IRQs can preempt the currently running thread more easily by
using a tailor made wait / signal process. In the wait function the IRQ
thread is taken of the run queue and put in a waiting state (note: the
thread is not put on any waiting queue). When the thread has to be woken
up the thread is reinserted in run queue and the
`THREAD_NEED_RESCHED_FLAG' is set on the currently running thread.

Signed-off-by: Michel Megens <dev@bietje.net>