sched: core: only free system stacks after `kill()'
Only free the memory of stacks that have been allocated by the threading
API when the thread was created. To check, a new thread flag has been
introduced: `THREAD_SYSTEM_STACK'. This bit is set by `thread_init()' if
it allocates a new stack region.

Once the scheduling core empties the kill queue, it will only free the
stack if `THREAD_SYSTEM_STACK' is set. The thread structure itself is
_never_ freed-up.

Signed-off-by: Michel Megens <dev@bietje.net>