sched: core: kill a thread remotely
Allow a thread to kill another thread. Previously it was only able to
kill a thread from within that thread. This can become a problem (and
lead to complicated code with flags) if a main thread wants to kill a
few helper threads. The main thread can now kill such a thread using:

	thread_destroy(thread_ptr);
	join(thread_ptr);

Thread destroy will flag a thread for destruction. The actual kill
operation will be executed the next time the scheduler attempts to run
the 'thread-to-be-killed'.

Signed-off-by: Michel Megens <dev@bietje.net>