python: vm: change standard thread prio to default
By default, the thread priority of the thread running the Python VM is
set to the the value of SCHED_DEFAULT_PRIO (which is 120). Running
higher priority my cause your system to clog up. It won't starve, as
long as a starvation free algorithm is used, but threads wont be ran as
often as they should be. Using the default priority gives the VM just
the attention it needs.

Signed-off-by: Michel Megens <dev@bietje.net>