sched: clock: add fix for non-sched configurations
Implement a fix to solve a linking error on ETA/OS configurations that
don't compile the scheduler.

Schedless configurations would run into a link error because
`sched_clock_tick()' will not be compiled. Fixed by adding an empty
implementation as inline function to `sched.h'. This version is only
used if the 'normal' one isn't compiled.

Signed-off-by: Michel Megens <dev@bietje.net>