sched: core: update the scheduling policy
Reschedule operations are dependant on the PREEMPT_NEED_RESCHED_FLAG and
the THREAD_NEED_RESCHED_FLAG. The latter is mostly for applications to
force a reschedule (e.g. using sleep or wait). This patch changes the
conditions on which the PREEMPT_NEED_RESCHED_FLAG is set:

	* the time slice of the current thread has expired;
	* the scheduling algorithm wants to preempt
	* the current thread has to much CPU time (SCHED_FAIR)

If SCHED_FAIR is enabled, the scheduling core will preempt the current
thread, if the *next* thread has less CPU time (disregarding the
algorithms decision to preempt or not).

If the *next* thread is the *idle thread*, the PREEMPT_NEED_RESCHED_FLAG
is discarded.

Signed-off-by: Michel Megens <dev@bietje.net>