sched: core: add next thread as argument to __schedule_prepare
Add a pointer to the `next' thread (i.e. the thread we are about to
switch to) as an argument to `__schedule_prepare()'. This way
`__schedule_prepare()' doesn't need to figure out what thread is next on
its own (which could cause weird bugs if the scheduler is updated in the
future).

Signed-off-by: Michel Megens <dev@bietje.net>