sched: core: move resched decisions to `reschedule_decide()'
Let `reschedule_decide()' be functionally responsible for the entire
reschedule decision. A reschedule will only be issued if:

	* the THREAD_NEED_RESCHED_FLAG is set or
	* PREEMPT_NEED_RESCHED_FLAG is set (iff next != thread_idle)

The values of the reschedule flags will be ignored if the current thread
is equal to the `next' thread. The reschedule decision can also be
expressed in the following logical expression:

	* p ∨ (v ∧ ¬q) && current_thread != next_thread

Where:
	* p is the value of the THREAD_NEED_RESCHED_FLAG
	* v is the value of the PREEMPT_NEED_RESCHED_FLAG
	* q is the value of the THREAD_IDLE FLAG of the next thread

Signed-off-by: Michel Megens <dev@bietje.net>